system-design/3.add_postgresql/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
passlib[bcrypt]
bcrypt
python-jose[cryptography]
python-multipart